+++ package.json
@@ -1,28 +1,31 @@
 {
-  "name": "next-project",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@apollo/client": "^3.11.8",
-    "graphql": "^16.9.0",
-    "next": "14.2.13",
-    "react": "^18",
-    "react-dom": "^18"
-  },
-  "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "eslint": "^8",
-    "eslint-config-next": "14.2.13",
-    "postcss": "^8",
-    "tailwindcss": "^3.4.1",
-    "typescript": "^5"
-  }
+    "name": "next-project",
+    "version": "0.1.0",
+    "private": true,
+    "scripts": {
+        "dev": "next dev",
+        "build": "next build",
+        "start": "next start",
+        "lint": "next lint",
+        "codegen": "graphql-codegen --config codegen.ts"
+    },
+    "dependencies": {
+        "@apollo/client": "^3.11.8",
+        "graphql": "^16.9.0",
+        "next": "14.2.13",
+        "react": "^18",
+        "react-dom": "^18"
+    },
+    "devDependencies": {
+        "@graphql-codegen/cli": "5.0.3",
+        "@types/node": "^20",
+        "@types/react": "^18",
+        "@types/react-dom": "^18",
+        "eslint": "^8",
+        "eslint-config-next": "14.2.13",
+        "postcss": "^8",
+        "tailwindcss": "^3.4.1",
+        "typescript": "^5.6.3",
+        "@graphql-codegen/client-preset": "4.4.0"
+    }
 }
