+++ src/app/boards/[boardId]/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import BoardsComponentDetail from '../../../components/boards-detail';
+
+export default function BoardsDetailPage() {
+    return <BoardsComponentDetail />;
+}
