+++ src/app/boards/new/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import BoardComponentWrite from '../../../next-project/src/components/boards-write';
+import BoardComponentWrite from '../../../components/boards-write';
 
 export default function BoardNewPage() {
     return <BoardComponentWrite isEdit={false} />;
