+++ src/app/boards/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import BoardsComponentList from '../../components/boards-list';
+
+export default function BoardList(props: any) {
+    return <BoardsComponentList />;
+}
