+++ src/assets/beach.png
