+++ src/assets/delete.png
