+++ src/assets/main-logo.png
