+++ src/components/boards-detail/queries.ts
@@ -1,0 +1,17 @@
+'use client';
+
+import { gql } from '@apollo/client';
+
+export const FETCH_BOARD = gql`
+    query fetchBoard($boardId: ID!) {
+        fetchBoard(boardId: $boardId) {
+            _id
+            writer
+            title
+            contents
+            youtubeUrl
+            likeCount
+            dislikeCount
+        }
+    }
+`;
