+++ src/components/boards-list/styles.module.css
@@ -118,6 +118,3 @@
     line-height: 20px;
     text-align: center;
 }
-
-.BoardListBoxDelete {
-}
